# server/docs/swagger.yaml
definitions:
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  user.FaUser:
    properties:
      B1total:
        type: number
      B1yest:
        type: number
      B2total:
        type: number
      B2yest:
        type: number
      B3total:
        type: number
      B3yest:
        type: number
      S1:
        type: number
      S2:
        type: number
      avatar:
        type: string
      bio:
        type: string
      birthday:
        type: string
      card1:
        type: string
      card2:
        type: string
      card3:
        type: string
      children:
        items:
          $ref: '#/definitions/user.FaUser'
        type: array
      createdAt:
        description: 创建时间
        type: string
      email:
        type: string
      faUserLevel:
        $ref: '#/definitions/user.FaUserLevel'
      gender:
        type: boolean
      groupId:
        type: integer
      gryj:
        type: integer
      id:
        description: 主键ID
        type: integer
      idcard:
        type: string
      isAgent:
        type: string
      isB:
        type: string
      isSim:
        type: string
      isSm:
        type: string
      jointime:
        type: integer
      level:
        type: integer
      loginfailure:
        type: boolean
      loginip:
        type: string
      logintime:
        type: integer
      mobile:
        type: string
      nickname:
        type: string
      password:
        type: string
      password2:
        type: string
      password3:
        type: string
      realname:
        type: string
      salt:
        type: string
      status:
        type: integer
      suan:
        type: string
      suan2:
        type: string
      sw1:
        type: integer
      sw2:
        type: integer
      sw3:
        type: integer
      sw4:
        type: integer
      sw5:
        type: integer
      sw6:
        type: integer
      tdnum:
        type: integer
      tdyj:
        type: integer
      tgno:
        type: string
      times:
        type: integer
      tjid:
        type: integer
      tjuser:
        type: string
      todayzt:
        type: integer
      tpath:
        type: string
      updatedAt:
        description: 更新时间
        type: string
      username:
        type: string
      viptime:
        type: integer
      wall1:
        type: number
      wall1freeze:
        type: number
      wall2:
        type: number
      wall2freeze:
        type: number
      wall3:
        type: number
      wall3freeze:
        type: number
      wall4:
        type: number
      wall4freeze:
        type: number
      wall5:
        type: number
      wall5freeze:
        type: number
      wall6:
        type: number
      wall6freeze:
        type: number
      wall7:
        type: number
      wall7freeze:
        type: number
      wall8:
        type: number
      wall8freeze:
        type: number
      wall9:
        type: number
      wall9freeze:
        type: number
      wall10:
        type: number
      wall10freeze:
        type: number
      win:
        type: string
      ztlevel:
        type: string
      ztnum:
        type: integer
    type: object
  user.FaUserLevel:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      n1:
        type: integer
      name:
        type: string
      rate:
        type: number
      tdnum:
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      ztnum:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /faUser/UserCharge:
    post:
      consumes:
      - application/json
      parameters:
      - in: query
        name: msg
        type: string
      - in: query
        name: price
        type: number
      - in: query
        name: ptype
        type: string
      - in: query
        name: typec
        type: string
      - in: query
        name: userid
        type: integer
      - in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 后台充值
      tags:
      - FaUser
  /faUser/UserTeam:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: msg
        type: string
      - in: query
        name: price
        type: number
      - in: query
        name: ptype
        type: string
      - in: query
        name: typec
        type: string
      - in: query
        name: userid
        type: integer
      - in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 后台查看团队
      tags:
      - FaUser
  /faUser/createFaUser:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建FaUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.FaUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建FaUser
      tags:
      - FaUser
  /faUser/deleteFaUser:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除FaUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.FaUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除FaUser
      tags:
      - FaUser
  /faUser/deleteFaUserByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除FaUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除FaUser
      tags:
      - FaUser
  /faUser/findFaUser:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: B1total
        type: number
      - in: query
        name: B1yest
        type: number
      - in: query
        name: B2total
        type: number
      - in: query
        name: B2yest
        type: number
      - in: query
        name: B3total
        type: number
      - in: query
        name: B3yest
        type: number
      - in: query
        name: S1
        type: number
      - in: query
        name: S2
        type: number
      - in: query
        name: avatar
        type: string
      - in: query
        name: bio
        type: string
      - in: query
        name: birthday
        type: string
      - in: query
        name: card1
        type: string
      - in: query
        name: card2
        type: string
      - in: query
        name: card3
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: email
        type: string
      - in: query
        name: gender
        type: boolean
      - in: query
        name: groupId
        type: integer
      - in: query
        name: gryj
        type: integer
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: idcard
        type: string
      - in: query
        name: isAgent
        type: string
      - in: query
        name: isB
        type: string
      - in: query
        name: isSim
        type: string
      - in: query
        name: isSm
        type: string
      - in: query
        name: jointime
        type: integer
      - in: query
        name: level
        type: integer
      - in: query
        name: loginfailure
        type: boolean
      - in: query
        name: loginip
        type: string
      - in: query
        name: logintime
        type: integer
      - in: query
        name: mobile
        type: string
      - in: query
        name: nickname
        type: string
      - in: query
        name: password
        type: string
      - in: query
        name: password2
        type: string
      - in: query
        name: password3
        type: string
      - in: query
        name: realname
        type: string
      - in: query
        name: salt
        type: string
      - in: query
        name: status
        type: integer
      - in: query
        name: suan
        type: string
      - in: query
        name: suan2
        type: string
      - in: query
        name: sw1
        type: integer
      - in: query
        name: sw2
        type: integer
      - in: query
        name: sw3
        type: integer
      - in: query
        name: sw4
        type: integer
      - in: query
        name: sw5
        type: integer
      - in: query
        name: sw6
        type: integer
      - in: query
        name: tdnum
        type: integer
      - in: query
        name: tdyj
        type: integer
      - in: query
        name: tgno
        type: string
      - in: query
        name: times
        type: integer
      - in: query
        name: tjid
        type: integer
      - in: query
        name: tjuser
        type: string
      - in: query
        name: todayzt
        type: integer
      - in: query
        name: tpath
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: username
        type: string
      - in: query
        name: viptime
        type: integer
      - in: query
        name: wall1
        type: number
      - in: query
        name: wall10
        type: number
      - in: query
        name: wall10freeze
        type: number
      - in: query
        name: wall1freeze
        type: number
      - in: query
        name: wall2
        type: number
      - in: query
        name: wall2freeze
        type: number
      - in: query
        name: wall3
        type: number
      - in: query
        name: wall3freeze
        type: number
      - in: query
        name: wall4
        type: number
      - in: query
        name: wall4freeze
        type: number
      - in: query
        name: wall5
        type: number
      - in: query
        name: wall5freeze
        type: number
      - in: query
        name: wall6
        type: number
      - in: query
        name: wall6freeze
        type: number
      - in: query
        name: wall7
        type: number
      - in: query
        name: wall7freeze
        type: number
      - in: query
        name: wall8
        type: number
      - in: query
        name: wall8freeze
        type: number
      - in: query
        name: wall9
        type: number
      - in: query
        name: wall9freeze
        type: number
      - in: query
        name: win
        type: string
      - in: query
        name: ztlevel
        type: string
      - in: query
        name: ztnum
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询FaUser
      tags:
      - FaUser
  /faUser/getFaUserList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: B1total
        type: number
      - in: query
        name: B1yest
        type: number
      - in: query
        name: B2total
        type: number
      - in: query
        name: B2yest
        type: number
      - in: query
        name: B3total
        type: number
      - in: query
        name: B3yest
        type: number
      - in: query
        name: S1
        type: number
      - in: query
        name: S2
        type: number
      - in: query
        name: avatar
        type: string
      - in: query
        name: bio
        type: string
      - in: query
        name: birthday
        type: string
      - in: query
        name: card1
        type: string
      - in: query
        name: card2
        type: string
      - in: query
        name: card3
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: email
        type: string
      - in: query
        name: endCreatedAt
        type: string
      - in: query
        name: gender
        type: boolean
      - in: query
        name: groupId
        type: integer
      - in: query
        name: gryj
        type: integer
      - description: 主键ID
        in: query
        name: id
        type: integer
      - in: query
        name: idcard
        type: string
      - in: query
        name: isAgent
        type: string
      - in: query
        name: isB
        type: string
      - in: query
        name: isSim
        type: string
      - in: query
        name: isSm
        type: string
      - in: query
        name: jointime
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: level
        type: integer
      - in: query
        name: loginfailure
        type: boolean
      - in: query
        name: loginip
        type: string
      - in: query
        name: logintime
        type: integer
      - in: query
        name: mobile
        type: string
      - in: query
        name: nickname
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: password
        type: string
      - in: query
        name: password2
        type: string
      - in: query
        name: password3
        type: string
      - in: query
        name: realname
        type: string
      - in: query
        name: salt
        type: string
      - in: query
        name: startCreatedAt
        type: string
      - in: query
        name: status
        type: integer
      - in: query
        name: suan
        type: string
      - in: query
        name: suan2
        type: string
      - in: query
        name: sw1
        type: integer
      - in: query
        name: sw2
        type: integer
      - in: query
        name: sw3
        type: integer
      - in: query
        name: sw4
        type: integer
      - in: query
        name: sw5
        type: integer
      - in: query
        name: sw6
        type: integer
      - in: query
        name: tdnum
        type: integer
      - in: query
        name: tdyj
        type: integer
      - in: query
        name: tgno
        type: string
      - in: query
        name: times
        type: integer
      - in: query
        name: tjid
        type: integer
      - in: query
        name: tjuser
        type: string
      - in: query
        name: todayzt
        type: integer
      - in: query
        name: tpath
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      - in: query
        name: username
        type: string
      - in: query
        name: viptime
        type: integer
      - in: query
        name: wall1
        type: number
      - in: query
        name: wall10
        type: number
      - in: query
        name: wall10freeze
        type: number
      - in: query
        name: wall1freeze
        type: number
      - in: query
        name: wall2
        type: number
      - in: query
        name: wall2freeze
        type: number
      - in: query
        name: wall3
        type: number
      - in: query
        name: wall3freeze
        type: number
      - in: query
        name: wall4
        type: number
      - in: query
        name: wall4freeze
        type: number
      - in: query
        name: wall5
        type: number
      - in: query
        name: wall5freeze
        type: number
      - in: query
        name: wall6
        type: number
      - in: query
        name: wall6freeze
        type: number
      - in: query
        name: wall7
        type: number
      - in: query
        name: wall7freeze
        type: number
      - in: query
        name: wall8
        type: number
      - in: query
        name: wall8freeze
        type: number
      - in: query
        name: wall9
        type: number
      - in: query
        name: wall9freeze
        type: number
      - in: query
        name: win
        type: string
      - in: query
        name: ztlevel
        type: string
      - in: query
        name: ztnum
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取FaUser列表
      tags:
      - FaUser
  /faUser/updateFaUser:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新FaUser
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/user.FaUser'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新FaUser
      tags:
      - FaUser
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"